+++ polytope_feature/engine/engine.py
@@ -6,6 +6,7 @@ from ..shapes import ConvexPolytope, Product
 
 from ..datacube.datacube_axis import UnsliceableDatacubeAxis
 from ..utility.list_tools import unique
+import math
 
 
 class Engine:
@@ -32,6 +33,18 @@ class Engine:
                 p.points[j][i] = mapper.to_float(mapper.parse(p.points[j][i]))
         # Remove duplicate points
         unique(p.points)
+
+    def remap_values(self, ax, value):
+        remapped_val = self.remapped_vals.get((value, ax.name), None)
+        if remapped_val is None:
+            remapped_val = value
+            if ax.is_cyclic:
+                remapped_val_interm = ax.remap([value, value])[0]
+                remapped_val = (remapped_val_interm[0] + remapped_val_interm[1]) / 2
+            if ax.can_round:
+                remapped_val = round(remapped_val, int(-math.log10(ax.tol)))
+            self.remapped_vals[(value, ax.name)] = remapped_val
+        return remapped_val
 
     def pre_process_polytopes(self, datacube, polytopes):
         for p in polytopes:
