+++ polytope_feature/engine/slicing_tools.py
@@ -1,0 +1,76 @@
+from copy import copy
+from itertools import chain
+
+import scipy.spatial
+
+from ..shapes import ConvexPolytope
+from ..utility.geometry import lerp
+from ..utility.list_tools import argmax, argmin
+
+
+def _find_intersects(polytope, slice_axis_idx, value):
+    intersects = []
+    # Find all points above and below slice axis
+    above_slice = [p for p in polytope.points if p[slice_axis_idx] >= value]
+    below_slice = [p for p in polytope.points if p[slice_axis_idx] <= value]
+
+    # Get the intersection of every pair above and below, this will create excess interior points
+    for a in above_slice:
+        for b in below_slice:
+            # edge is incident with slice plane, don't need these points
+            if a[slice_axis_idx] == b[slice_axis_idx]:
+                intersects.append(b)
+                continue
+
+            # Linearly interpolate all coordinates of two points (a,b) of the polytope
+            interp_coeff = (value - b[slice_axis_idx]) / (a[slice_axis_idx] - b[slice_axis_idx])
+            intersect = lerp(a, b, interp_coeff)
+            intersects.append(intersect)
+    return intersects
+
+
+def _reduce_dimension(intersects, slice_axis_idx):
+    temp_intersects = []
+    for point in intersects:
+        point = [p for i, p in enumerate(point) if i != slice_axis_idx]
+        temp_intersects.append(point)
+    return temp_intersects
+
+
+def slice(polytope: ConvexPolytope, axis, value, slice_axis_idx):
+    if polytope.is_flat:
+        if value in chain(*polytope.points):
+            intersects = [[value]]
+        else:
+            return None
+    else:
+        intersects = _find_intersects(polytope, slice_axis_idx, value)
+
+    if len(intersects) == 0:
+        return None
+
+    # Reduce dimension of intersection points, removing slice axis
+    intersects = _reduce_dimension(intersects, slice_axis_idx)
+
+    axes = copy(polytope._axes)
+    axes.remove(axis)
+
+    if len(intersects) < len(intersects[0]) + 1:
+        return ConvexPolytope(axes, intersects)
+    # Compute convex hull (removing interior points)
+    if len(intersects[0]) == 0:
+        return None
+    elif len(intersects[0]) == 1:  # qhull doesn't like 1D, do it ourselves
+        amin = argmin(intersects)
+        amax = argmax(intersects)
+        vertices = [amin, amax]
+    else:
+        try:
+            hull = scipy.spatial.ConvexHull(intersects)
+            vertices = hull.vertices
+
+        except scipy.spatial.qhull.QhullError as e:
+            if "less than" or "flat" in str(e):
+                return ConvexPolytope(axes, intersects)
+    # Sliced result is simply the convex hull
+    return ConvexPolytope(axes, [intersects[i] for i in vertices])
