+++ polytope_feature/utility/combinatorics.py
@@ -2,7 +2,7 @@ import itertools
 from collections import Counter
 from typing import List
 
-from ..shapes import ConvexPolytope
+from ..shapes import ConvexPolytope, Product
 from .exceptions import AxisNotFoundError, AxisOverdefinedError, AxisUnderdefinedError
 
 
@@ -46,3 +46,29 @@ def validate_axes(actual_axes, test_axes):
         raise AxisNotFoundError(ax)
 
     return True
+
+
+def find_polytopes_on_axis(axis_name, polytopes):
+    polytopes_on_axis = []
+    for poly in polytopes:
+        if axis_name in poly._axes:
+            polytopes_on_axis.append(poly)
+    return polytopes_on_axis
+
+
+def find_polytope_combinations(c):
+    new_c = []
+    for combi in c:
+        if isinstance(combi, list):
+            new_c.extend(combi)
+        else:
+            new_c.append(combi)
+    # NOTE TODO: here some of the polys in new_c can be a Product shape instead of a ConvexPolytope
+    # -> need to go through the polytopes in new_c and replace the Products with their sub-ConvexPolytopes
+    final_polys = []
+    for poly in new_c:
+        if isinstance(poly, Product):
+            final_polys.extend(poly.polytope())
+        else:
+            final_polys.append(poly)
+    return final_polys
